fix #455 -- reduce precision to 1 second instead of 1 millisecond

+++ opentripplanner-routing/src/main/java/org/opentripplanner/routing/services/RoutingService.java
@@ -46,5 +46,5 @@ public interface RoutingService {
      * @return
      */
     public GraphPath route(Vertex fromVertex, Vertex toVertex, List<Vertex> intermediateVertices,
-            long time, TraverseOptions options);
+            int time, TraverseOptions options);
 }
